move stumptown to scraper folder, fix all the scraper problems
