+++ scrapers/stumptown.py
@@ -1,0 +1,67 @@
+from bs4 import BeautifulSoup
+import requests
+from models import Coffee
+import logging
+import re
+
+# scraping stumptown
+def scrape():
+    roaster = 'Stumptown'
+    stumptown = 'https://www.stumptowncoffee.com/coffee'
+
+    r = requests.get(stumptown)
+    soup = BeautifulSoup(r.content, "html.parser")
+    # class="product-grid _link"
+    coffees_for_sale = soup.find_all('a', {'class':'product-grid _link'})
+
+    for items in coffees_for_sale:
+        url = items['href']
+        if not 'trio' in url:
+            name,price,description,notes,region,status,size = [""]*7
+            product_url = 'https://www.stumptowncoffee.com'+url
+            logging.info("Getting url: {}".format(url))
+            r = requests.get(product_url)
+            coffee_soup = BeautifulSoup(r.content)
+            # product name h1 class="product _title -desktop theme-color js-pdp-title"
+            name = coffee_soup.h1.string.strip()
+            try:
+                price = int(coffee_soup.find_all('span',{'class':'js-pdp-price'})[0].string)
+            except IndexError as e:
+                logging.warn("Error while getting price for {} : {}".format(name, e))
+            # div class="product _description
+            description = coffee_soup.find('div', {'class':'product _description'}).p.string
+            try:
+                notes = coffee_soup.h3.string
+            except:
+                pass
+
+            try:
+                region = coffee_soup.find_all('h4')[1].span.string.strip()[8:]
+            except:
+                pass
+
+            if coffee_soup.h6:
+                status = 'Sold Out'
+            else:
+                status = 'Available'
+
+            # size in ounces
+            try:
+                size = int(re.findall('\d+', coffee_soup.find('div', {'class':'product _specs'}).find_all('p')[1].string)[0])
+            except Exception as e:
+                logging.warn("Error while getting size for {} : {}".format(name, e))
+            coffee_data = {'name': name, 'roaster': roaster, 'description': description, 'price': price, 'notes': notes, 'region': region, 'status': status, 'product_page': product_url, 'size': size}
+            print(coffee_data)
+            old_coffees = Coffee.query(Coffee.name == coffee_data['name'], Coffee.roaster == coffee_data['roaster'], Coffee.region == coffee_data['region']).fetch()
+            if old_coffees:
+                if len(old_coffees) > 1:
+                    logging.warning("Query for coffee name: {}, roaster: {}, region: {} returned {} results. Result are {}".format(coffee_data['name'], coffee_data['roaster'], coffee_data['region'], len(old_coffees), old_coffees))
+                for key, value in coffee_data.iteritems():
+                    setattr(old_coffees[0], key, value)
+                old_coffees[0].put
+            else: 
+                coffee = Coffee(**coffee_data)
+                coffee.put()
+
+
+
